Continue transaction implementation
