+++ chains/Cargo.toml
@@ -10,6 +10,7 @@ edition = "2018"
 farcaster_core = { path = "../core" }
 strict_encoding = "1.2.1"
 hex = "0.4.3"
+thiserror = "1.0.24"
 
 # blockchain specific
 bitcoin = "0.26.0"
